Changing ORWaveform: value method to take unsigned long instead of unsigned short.
Updating other classes that have this also as a dependency.

git-svn-id: svn://orca.physics.unc.edu/Orca@6119 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Data Processing/Histogram Objects/BaseClasses/ORDataController.h
@@ -80,7 +80,7 @@
 
 @interface NSObject (ORDataController_Cat)
 - (int)  numberBins;
-- (long) value:(unsigned short)aChan;
+- (long) value:(unsigned long)aChan;
 @end
