+++ Source/Objects/Data Processing/Histogram Objects/BaseClasses/ORDataController.m
@@ -463,7 +463,7 @@ int windowSort(id w1, id w2, void *context) { return [[w2 title] compare:[w1 tit
     return 0;
 }
 
-- (long) value:(unsigned short)aChan;
+- (long) value:(unsigned long)aChan;
 {
     return 0;
 }
